Make search compatible with authorization component
The Search plugin does a redirect in its startup method. Because this event happens before your controller action and aborts the request, the skipAuthorization() and authorize() calls in your controller actions are never reached on the redirect response. cakephp/authorization#128

Moving the logic from the startup() to beforeRender() solves the issue.
#258 #265